Freebase refactoring: Freebase implements Iterable<FreebaseNode> (#194)
Revised Freebase API: Freebase now implements Iterable<FreebaseNode>. A FreebaseNode is a node in the Freebase knowledge graph. The API allows us to iterate over nodes - each node contains all the outgoing edges, i.e., all facts about that object. FreebaseNode replaces what we were calling ObjectTriples. cf. #188
