+++ src/main/java/io/anserini/kg/freebase/IndexNodes.java
@@ -7,6 +7,10 @@ import org.apache.lucene.analysis.en.EnglishAnalyzer;
 import org.apache.lucene.codecs.lucene50.Lucene50StoredFieldsFormat;
 import org.apache.lucene.codecs.lucene62.Lucene62Codec;
 import org.apache.lucene.document.Document;
+import org.apache.lucene.document.Field;
+import org.apache.lucene.document.StoredField;
+import org.apache.lucene.document.StringField;
+import org.apache.lucene.document.TextField;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.store.Directory;
@@ -17,24 +21,25 @@ import org.kohsuke.args4j.Option;
 import org.kohsuke.args4j.OptionHandlerFilter;
 import org.kohsuke.args4j.ParserProperties;
 import org.kohsuke.args4j.spi.StringArrayOptionHandler;
+import org.openrdf.rio.ntriples.NTriplesUtil;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Builds a triples lookup index from a Freebase dump in N-Triples RDF format. Each
- * {@link ObjectTriples} object, which represents a group of triples that share the same subject,
+ * {@link FreebaseNode} object, which represents a group of triples that share the same subject,
  * is treated as a Lucene "document". This class builds an index primarily for lookup by
  * <code>mid</code>.
  */
-public class IndexObjectTriples {
-  private static final Logger LOG = LogManager.getLogger(IndexObjectTriples.class);
+public class IndexNodes {
+  private static final Logger LOG = LogManager.getLogger(IndexNodes.class);
 
   public static final class Args {
     // Required arguments
@@ -70,7 +75,7 @@ public class IndexObjectTriples {
    * @param args program arguments
    * @throws Exception
    */
-  public IndexObjectTriples(Args args) throws Exception {
+  public IndexNodes(Args args) throws Exception {
 
     // Copy arguments
     this.args = args;
@@ -138,14 +143,13 @@ public class IndexObjectTriples {
    * @throws IOException on error
    */
   private void index(IndexWriter writer, Path inputFile) throws IOException {
-    ObjectTriplesLuceneDocumentGenerator transformer = new ObjectTriplesLuceneDocumentGenerator();
+    NodeLuceneDocumentGenerator transformer = new NodeLuceneDocumentGenerator();
     transformer.config(args);
     transformer.setCounters(counters);
 
     int cnt = 0;
-    ObjectTriplesIterator iter = new ObjectTriplesIterator(inputFile);
-    while (iter.hasNext()) {
-      ObjectTriples d = iter.next();
+    Freebase freebase = new Freebase(inputFile);
+    for (FreebaseNode d : freebase) {
       Document doc = transformer.createDocument(d);
 
       if (doc != null) {
@@ -157,14 +161,9 @@ public class IndexObjectTriples {
       if (cnt % 100000 == 0) {
         LOG.debug("Number of indexed entity document: {}", cnt);
       }
-
-      d = null;
-      doc = null;
     }
 
-    iter.close();
-    LOG.info(inputFile.getParent().getFileName().toString() + File.separator +
-            inputFile.getFileName().toString() + ": " + cnt + " docs added.");
+    LOG.info(inputFile.getFileName().toString() + ": " + cnt + " docs added.");
     counters.indexedDocuments.addAndGet(cnt);
   }
 
@@ -178,11 +177,149 @@ public class IndexObjectTriples {
     } catch (CmdLineException e) {
       System.err.println(e.getMessage());
       parser.printUsage(System.err);
-      System.err.println("Example command: "+ IndexObjectTriples.class.getSimpleName() +
+      System.err.println("Example command: "+ IndexNodes.class.getSimpleName() +
               parser.printExample(OptionHandlerFilter.REQUIRED));
       return;
     }
 
-    new IndexObjectTriples(indexRDFCollectionArgs).run();
+    new IndexNodes(indexRDFCollectionArgs).run();
+  }
+
+  public static class NodeLuceneDocumentGenerator {
+    public static final String FIELD_SUBJECT = "subject";
+
+    // RDF object predicate types
+    static final String VALUE_TYPE_URI = "URI";
+    static final String VALUE_TYPE_STRING = "STRING";
+    static final String VALUE_TYPE_TEXT = "TEXT";
+    static final String VALUE_TYPE_OTHER = "OTHER";
+
+    protected IndexNodes.Counters counters;
+    protected IndexNodes.Args args;
+
+    public void config(IndexNodes.Args args) {
+      this.args = args;
+    }
+
+    public void setCounters(IndexNodes.Counters counters) {
+      this.counters = counters;
+    }
+
+    public Document createDocument(FreebaseNode src) {
+      // Convert the triple doc to lucene doc
+      Document doc = new Document();
+
+      // Index subject as a StringField to allow searching
+      Field subjectField = new StringField(FIELD_SUBJECT,
+          cleanUri(src.getSubject()),
+          Field.Store.YES);
+      doc.add(subjectField);
+
+      // Iterate over predicates and object values
+      for (Map.Entry<String, List<String>> entry : src.getPredicateValues().entrySet()) {
+        String predicate = cleanUri(entry.getKey());
+        List<String> values = entry.getValue();
+
+        for (String value : values) {
+          String valueType = getObjectType(value);
+          value = normalizeObjectValue(value);
+          if (isIndexedPredicate(predicate)) {
+            if (valueType.equals(VALUE_TYPE_URI)) {
+              // Always index URIs using StringField
+              doc.add(new StringField(predicate, value, Field.Store.YES));
+            } else {
+              // Just store the predicate in a stored field, no index
+              doc.add(new TextField(predicate, value, Field.Store.YES));
+            }
+          } else {
+            // Just add the predicate as a stored field, no index on it
+            doc.add(new StoredField(predicate, value));
+          }
+        }
+      }
+
+      src.clear();
+      return doc;
+    }
+
+    /**
+     * Removes '<', '>' if they exist, lower case
+     * <p>
+     * TODO - replace ':' with '_' because query parser doesn't like it
+     *
+     * @param uri
+     * @return
+     */
+    public static String cleanUri(String uri) {
+      if (uri.charAt(0) == '<')
+        return uri.substring(1, uri.length() - 1).toLowerCase();
+      else
+        return uri;
+    }
+
+    /**
+     * Figures out the type of the value of an object in a triple
+     *
+     * @param objectValue object value
+     * @return uri, string, text, or other
+     */
+    public static String getObjectType(String objectValue) {
+      // Determine the type of this N-Triples `value'.
+      char first = objectValue.charAt(0);
+      switch (first) {
+        case '<':
+          return VALUE_TYPE_URI;
+        case '"':
+          if (objectValue.charAt(objectValue.length() - 1) == '"')
+            return VALUE_TYPE_STRING;
+          else
+            return VALUE_TYPE_TEXT;
+        default:
+          return VALUE_TYPE_OTHER;
+      }
+    }
+
+    /**
+     * Check if the predicate should be indexed
+     *
+     * @param predicate the predicate to check
+     * @return true if the user specified it to be indexed, false otherwise.
+     */
+    boolean isIndexedPredicate(String predicate) {
+      return args.predicatesToIndex != null && args.predicatesToIndex.contains(predicate);
+    }
+
+    /**
+     * Do nothing for strings
+     */
+    public static String normalizeStringValue(String value) {
+      return value;
+    }
+
+    /**
+     * Un-escape strings
+     */
+    public static String normalizeTextValue(String value) {
+      return NTriplesUtil.unescapeString(value);
+    }
+
+    /**
+     * Normalize Object object value
+     *
+     * @param objectValue
+     * @return
+     */
+    public static String normalizeObjectValue(String objectValue) {
+      // Normalize a `objectValue' depending on its type.
+      String type = getObjectType(objectValue);
+      if (type.equals(VALUE_TYPE_URI))
+        return cleanUri(objectValue);
+      else if (type.equals(VALUE_TYPE_STRING))
+        return normalizeStringValue(objectValue);
+      else if (type.equals(VALUE_TYPE_TEXT))
+        return normalizeTextValue(objectValue);
+      else
+        return objectValue;
+    }
   }
 }
